+++ models/elf_image.go
@@ -58,7 +58,7 @@ type ElfImage struct {
 
 	// size
 	// Required: true
-	Size *float64 `json:"size"`
+	Size *int64 `json:"size"`
 
 	// upload task
 	UploadTask *NestedUploadTask `json:"upload_task,omitempty"`
